+++ packages/forma-36-react-components/src/index.ts
@@ -77,7 +77,27 @@ export type {
 } from '@contentful/f36-button';
 export { TextLink } from '@contentful/f36-text-link';
 export type { TextLinkProps } from '@contentful/f36-text-link';
-export * from './components/Accordion';
+export type {
+  SkeletonBodyTextProps,
+  SkeletonContainerProps,
+  SkeletonDisplayTextProps,
+  SkeletonImageProps,
+  SkeletonRowProps,
+  SkeletonTextProps,
+} from '@contentful/f36-skeleton';
+export {
+  SkeletonBodyText,
+  SkeletonContainer,
+  SkeletonDisplayText,
+  SkeletonImage,
+  SkeletonText,
+  SkeletonRow,
+} from '@contentful/f36-skeleton';
+export { Accordion, AccordionItem } from '@contentful/f36-accordion';
+export type {
+  AccordionProps,
+  AccordionItemProps,
+} from '@contentful/f36-accordion';
 export * from './components/Asset';
 export * from './components/Autocomplete';
 export * from './components/Card';
@@ -92,7 +112,6 @@ export * from './components/Modal';
 export * from './components/Notification';
 export * from './components/Select';
 export * from './components/SelectField';
-export * from './components/Skeleton';
 export * from './components/Switch';
 export * from './components/Tabs';
 export * from './components/Textarea';